requirements.txt
langgraph 
langchain 
openai 
pydantic
langsmith
python-dotenv
langchain-openai
streamlit
fastapi
uvicorn[standard]